[refac] Refactor AITaskService and integrate with StudentRecordAIFacade and SSEChannelManager
